--- api-v1-jackson/src/main/kotlin/ru/otus/otuskotlin/api/v1/ApiV1Mapper.kt
@@ -1,8 +1,0 @@
-package ru.otus.otuskotlin.api.v1
-
-import com.fasterxml.jackson.databind.ObjectMapper
-
-val apiV1Mapper = ObjectMapper().apply {
-//    configure(DeserializationFeature.FAIL_ON_UNKNOWN_PROPERTIES, false)
-//    setSerializationInclusion(JsonInclude.Include.NON_NULL)
-}
