--- common/src/main/kotlin/ru/otus/otuskotlin/common/models/IctrlCommand.kt
@@ -1,10 +1,0 @@
-package ru.otus.otuskotlin.common.models
-
-enum class IctrlCommand {
-    NONE,
-    CREATE,
-    READ,
-    UPDATE,
-    DELETE,
-    SEARCH,
-}
