--- common/src/main/kotlin/ru/otus/otuskotlin/common/models/IctrlCommodityId.kt
@@ -1,10 +1,0 @@
-package ru.otus.otuskotlin.common.models
-
-@JvmInline
-value class IctrlCommodityId(private val id: String) {
-    fun asString() = id
-
-    companion object {
-        val NONE = IctrlCommodityId("")
-    }
-}
