--- common/src/main/kotlin/ru/otus/otuskotlin/common/models/IctrlCommodityPermissionClient.kt
@@ -1,10 +1,0 @@
-package ru.otus.otuskotlin.common.models
-
-enum class IctrlCommodityPermissionClient {
-    READ,
-    UPDATE,
-    DELETE,
-    MAKE_VISIBLE_PUBLIC,
-    MAKE_VISIBLE_GROUP,
-    MAKE_VISIBLE_OWNER,
-}
