--- common/src/main/kotlin/ru/otus/otuskotlin/common/models/IctrlCommodityType.kt
@@ -1,8 +1,0 @@
-package ru.otus.otuskotlin.common.models
-
-enum class IctrlCommodityType {
-    TUBE_LINE_PART,
-    FASTENER_PART,
-    COATING_MATERIAL,
-    NONE,
-}
