--- common/src/main/kotlin/ru/otus/otuskotlin/common/models/IctrlState.kt
@@ -1,7 +1,0 @@
-package ru.otus.otuskotlin.common.models
-
-enum class IctrlState {
-    NONE,
-    RUNNING,
-    FAILING,
-}
