--- common/src/main/kotlin/ru/otus/otuskotlin/common/models/IctrlUserId.kt
@@ -1,10 +1,0 @@
-package ru.otus.otuskotlin.common.models
-
-@JvmInline
-value class IctrlUserId(private val id: String) {
-    fun asString() = id
-
-    companion object {
-        val NONE = IctrlUserId("")
-    }
-}
