+++ common/src/main/kotlin/ru/otus/otuskotlin/incomingControl/common/models/IctrlCommand.kt
@@ -1,0 +1,10 @@
+package ru.otus.otuskotlin.incomingControl.common.models
+
+enum class IctrlCommand {
+    NONE,
+    CREATE,
+    READ,
+    UPDATE,
+    DELETE,
+    SEARCH,
+}
