+++ common/src/main/kotlin/ru/otus/otuskotlin/incomingControl/common/models/IctrlCommodityFilter.kt
@@ -1,0 +1,6 @@
+package ru.otus.otuskotlin.incomingControl.common.models
+
+data class IctrlCommodityFilter(
+    var searchString: String = "",
+    var ownerId: IctrlUserId = IctrlUserId.NONE,
+)
