+++ common/src/main/kotlin/ru/otus/otuskotlin/incomingControl/common/models/IctrlCommodityPermissionClient.kt
@@ -1,0 +1,10 @@
+package ru.otus.otuskotlin.incomingControl.common.models
+
+enum class IctrlCommodityPermissionClient {
+    READ,
+    UPDATE,
+    DELETE,
+    MAKE_VISIBLE_PUBLIC,
+    MAKE_VISIBLE_GROUP,
+    MAKE_VISIBLE_OWNER,
+}
