+++ common/src/main/kotlin/ru/otus/otuskotlin/incomingControl/common/models/IctrlError.kt
@@ -1,0 +1,9 @@
+package ru.otus.otuskotlin.incomingControl.common.models
+
+data class IctrlError(
+    val code: String = "",
+    val group: String = "",
+    val field: String = "",
+    val message: String = "",
+    val exception: Throwable? = null,
+)
