+++ common/src/main/kotlin/ru/otus/otuskotlin/incomingControl/common/models/IctrlProductId.kt
@@ -1,0 +1,10 @@
+package ru.otus.otuskotlin.incomingControl.common.models
+
+@JvmInline
+value class IctrlProductId(private val id: String) {
+    fun asString() = id
+
+    companion object {
+        val NONE = IctrlProductId("")
+    }
+}
