+++ common/src/main/kotlin/ru/otus/otuskotlin/incomingControl/common/models/IctrlState.kt
@@ -1,0 +1,7 @@
+package ru.otus.otuskotlin.incomingControl.common.models
+
+enum class IctrlState {
+    NONE,
+    RUNNING,
+    FAILING,
+}
