--- mappers-v1/src/main/kotlin/ru/otus/otuskotlin/mappers/v1/exceptions/UnknownRequestClass.kt
@@ -1,4 +1,0 @@
-package ru.otus.otuskotlin.mappers.v1.exceptions
-
-class UnknownRequestClass(clazz: Class<*>) :
-    RuntimeException("Class $clazz cannot be mapped to ru.otus.otuskotlin.incomingcontrol.common.IctrlContext")
